Moved local data sources to data folder
