+++ composables/useData.js
@@ -1,4 +1,4 @@
-import data from "~/data.json";
+import data from "~/data/data.json";
 
 export const useData = () => {
   return useState("data", () => data);
